Write better tests
