+++ hardhat.config.js
@@ -28,6 +28,9 @@ module.exports = {
       },
     },
   },
+  mocha: {
+    parallel: true,
+  },
   contractSizer: {
     runOnCompile: true,
   },
